Response to CR 1-14
