+++ lms/djangoapps/shoppingcart/tests/test_models.py
@@ -18,12 +18,10 @@ from xmodule.modulestore.tests.factories import CourseFactory
 from courseware.tests.tests import TEST_DATA_MONGO_MODULESTORE
 from shoppingcart.models import (Order, OrderItem, CertificateItem, InvalidCartItem, PaidCourseRegistration,
                                  OrderItemSubclassPK, PaidCourseRegistrationAnnotation)
-from shoppingcart.views import initialize_report, REPORT_TYPES
-from shoppingcart.reports import ItemizedPurchaseReport, CertificateStatusReport, UniversityRevenueShareReport, RefundReport
 from student.tests.factories import UserFactory
 from student.models import CourseEnrollment
 from course_modes.models import CourseMode
-from shoppingcart.exceptions import PurchasedCallbackException, ReportTypeDoesNotExistException
+from shoppingcart.exceptions import PurchasedCallbackException
 import pytz
 import datetime
 
@@ -323,101 +321,6 @@ class PaidCourseRegistrationTest(ModuleStoreTestCase):
         with self.assertRaises(PurchasedCallbackException):
             reg1.purchased_callback()
         self.assertFalse(CourseEnrollment.is_enrolled(self.user, self.course_id))
-
-
-@override_settings(MODULESTORE=TEST_DATA_MONGO_MODULESTORE)
-class ItemizedPurchaseReportTest(ModuleStoreTestCase):
-    """
-    Tests for the models used to generate itemized purchase reports
-    """
-    FIVE_MINS = datetime.timedelta(minutes=5)
-    TEST_ANNOTATION = u'Ba\xfc\u5305'
-
-    def setUp(self):
-        self.user = UserFactory.create()
-        self.course_id = "MITx/999/Robot_Super_Course"
-        self.cost = 40
-        self.course = CourseFactory.create(org='MITx', number='999', display_name=u'Robot Super Course')
-        course_mode = CourseMode(course_id=self.course_id,
-                                 mode_slug="honor",
-                                 mode_display_name="honor cert",
-                                 min_price=self.cost)
-        course_mode.save()
-        course_mode2 = CourseMode(course_id=self.course_id,
-                                  mode_slug="verified",
-                                  mode_display_name="verified cert",
-                                  min_price=self.cost)
-        course_mode2.save()
-        self.annotation = PaidCourseRegistrationAnnotation(course_id=self.course_id, annotation=self.TEST_ANNOTATION)
-        self.annotation.save()
-        self.cart = Order.get_cart_for_user(self.user)
-        self.reg = PaidCourseRegistration.add_to_order(self.cart, self.course_id)
-        self.cert_item = CertificateItem.add_to_order(self.cart, self.course_id, self.cost, 'verified')
-        self.cart.purchase()
-        self.now = datetime.datetime.now(pytz.UTC)
-
-        paid_reg = PaidCourseRegistration.objects.get(course_id=self.course_id, user=self.user)
-        paid_reg.fulfilled_time = self.now
-        paid_reg.refund_requested_time = self.now
-        paid_reg.save()
-
-        cert = CertificateItem.objects.get(course_id=self.course_id, user=self.user)
-        cert.fulfilled_time = self.now
-        cert.refund_requested_time = self.now
-        cert.save()
-
-        self.CORRECT_CSV = dedent("""
-            Purchase Time,Order ID,Status,Quantity,Unit Cost,Total Cost,Currency,Description,Comments
-            {time_str},1,purchased,1,40,40,usd,Registration for Course: Robot Super Course,Ba\xc3\xbc\xe5\x8c\x85
-            {time_str},1,purchased,1,40,40,usd,"Certificate of Achievement, verified cert for course Robot Super Course",
-            """.format(time_str=str(self.now)))
-
-    def test_purchased_items_btw_dates(self):
-        report = initialize_report("itemized_purchase_report", self.now - self.FIVE_MINS, self.now + self.FIVE_MINS)
-        purchases = report.rows()
-
-        # since there's not many purchases, just run through the generator to make sure we've got the right number
-        num_purchases = 0
-        for item in purchases:
-            num_purchases += 1
-        self.assertEqual(num_purchases, 2)
-        #self.assertIn(self.reg.orderitem_ptr, purchases)
-        #self.assertIn(self.cert_item.orderitem_ptr, purchases)
-
-        report = initialize_report("itemized_purchase_report", self.now + self.FIVE_MINS, self.now + self.FIVE_MINS + self.FIVE_MINS)
-        no_purchases = report.rows()
-
-        num_purchases = 0
-        for item in no_purchases:
-            num_purchases +=1
-        self.assertEqual(num_purchases, 0)
-
-    def test_purchased_csv(self):
-        """
-        Tests that a generated purchase report CSV is as we expect
-        """
-        report = initialize_report("itemized_purchase_report", self.now - self.FIVE_MINS, self.now + self.FIVE_MINS)
-        csv_file = StringIO.StringIO()
-        report.write_csv(csv_file)
-        csv = csv_file.getvalue()
-        csv_file.close()
-        # Using excel mode csv, which automatically ends lines with \r\n, so need to convert to \n
-        self.assertEqual(csv.replace('\r\n', '\n').strip(), self.CORRECT_CSV.strip())
-
-    def test_csv_report_no_annotation(self):
-        """
-        Fill in gap in test coverage.  csv_report_comments for PaidCourseRegistration instance with no
-        matching annotation
-        """
-        # delete the matching annotation
-        self.annotation.delete()
-        self.assertEqual(u"", self.reg.csv_report_comments)
-
-    def test_paidcourseregistrationannotation_unicode(self):
-        """
-        Fill in gap in test coverage.  __unicode__ method of PaidCourseRegistrationAnnotation
-        """
-        self.assertEqual(unicode(self.annotation), u'{} : {}'.format(self.course_id, self.TEST_ANNOTATION))
 
 
 @override_settings(MODULESTORE=TEST_DATA_MONGO_MODULESTORE)
