+++ lms/djangoapps/shoppingcart/tests/test_views.py
@@ -21,7 +21,6 @@ from course_modes.models import CourseMode
 from edxmako.shortcuts import render_to_response
 from shoppingcart.processors import render_purchase_form_html
 from mock import patch, Mock, sentinel
-from shoppingcart.reports import ItemizedPurchaseReport
 from shoppingcart.views import initialize_report
