+++ lms/templates/shoppingcart/download_report.html
@@ -1,5 +1,6 @@
 <%! from django.utils.translation import ugettext as _ %>
 <%! from django.core.urlresolvers import reverse %>
+<%! from django.conf import settings %>
 <%inherit file="../main.html" />
 
 <%block name="title"><title>${_("Download CSV Reports")}</title></%block>
@@ -19,21 +20,35 @@
     <label for="end_date">${_("End Date: ")}</label>
     <input id="end_date" type="text" value="${end_date}" name="end_date"/>
     <br/>
+
+    %if "itemized_purchase_report" in settings.FEATURES['ENABLED_PAYMENT_REPORTS']:
     <button type = "submit" name="requested_report" value="itemized_purchase_report">Itemized Purchase Report</button>
     <br/>
-    <button type = "submit" name="requested_report" value="refund_report">Refund Report</button>
+    %endif
 
-    <br/><br/>
+    %if "refund_report" in settings.FEATURES['ENABLED_PAYMENT_REPORTS']:
+    <button type = "submit" name="requested_report" value="refund_report">Refund Report</button>
+    <br/>
+    %endif
+
+    <br/>
+
     <p>${_("These reports are delimited alphabetically by university name.  i.e., generating a report with 'Start Letter' A and 'End Letter' C will generate reports for all universities starting with A, B, and C.")}</p>
     <label for="start_letter">${_("Start Letter: ")}</label>
     <input id="start_letter" type="text" value="${start_letter}" name="start_letter"/>
     <label for="end_letter">${_("End Letter: ")}</label>
     <input id="end_letter" type="text" value="${end_letter}" name="end_letter"/>
     <input type="hidden" name="csrfmiddlewaretoken" value="${csrf_token}" />
-
     <br/>
+
+    %if "university_revenue_share" in settings.FEATURES['ENABLED_PAYMENT_REPORTS']:
     <button type = "submit" name="requested_report" value="university_revenue_share">University Revenue Share</button>
     <br/>
+    %endif
+
+    %if "certificate_status" in settings.FEATURES['ENABLED_PAYMENT_REPORTS']:
     <button type="submit" name="requested_report" value="certificate_status">Certiciate Status</button>
+    <br/>
+    %endif
   </form>
 </section>
